# etcd: kvdb fixes