--- a/kvdb/debug.go
+++ b/kvdb/debug.go
@@ -1,0 +1,8 @@
+// +build dev
+
+package kvdb
+
+const (
+	// Switch on extra debug code.
+	etcdDebug = true
+)