--- a/kvdb/etcd/debug.go
+++ b/kvdb/etcd/debug.go
@@ -1,0 +1,8 @@
+// +build dev
+
+package etcd
+
+const (
+	// Switch on extra debug code.
+	etcdDebug = true
+)