--- a/kvdb/etcd/nodebug.go
+++ b/kvdb/etcd/nodebug.go
@@ -1,0 +1,8 @@
+// +build !dev
+
+package etcd
+
+const (
+	// Switch off extra debug code.
+	etcdDebug = false
+)