--- a/kvdb/etcd_test.go
+++ b/kvdb/etcd_test.go
@@ -19,6 +19,7 @@ var (
 func TestEtcd(t *testing.T) {
 	tests := []struct {
 		name       string
+		debugOnly  bool
 		test       func(*testing.T, walletdb.DB)
 		expectedDb map[string]string
 	}{
@@ -103,8 +104,9 @@ func TestEtcd(t *testing.T) {
 			test: testBucketSequence,
 		},
 		{
-			name: "key clash",
-			test: testKeyClash,
+			name:      "key clash",
+			debugOnly: true,
+			test:      testKeyClash,
 			expectedDb: map[string]string{
 				bkey("apple"):           bval("apple"),
 				bkey("apple", "banana"): bval("apple", "banana"),
@@ -136,6 +138,10 @@ func TestEtcd(t *testing.T) {
 
 	for _, test := range tests {
 		test := test
+
+		if test.debugOnly && !etcdDebug {
+			continue
+		}
 
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()