--- a/kvdb/nodebug.go
+++ b/kvdb/nodebug.go
@@ -1,0 +1,8 @@
+// +build !dev
+
+package kvdb
+
+const (
+	// Switch off extra debug code.
+	etcdDebug = false
+)