This PR fixes a bug in the cursor where `Delete` would step the cursor to the next element after the deleted one which is incompatible with bbolt. 

Furthermore we remove an assertion that would prevent us from creating a value with a key that already is used for a bucket.  While enforcing this rule is helpful it's impractical as we don't have such buckets and values. Removing this assertion improves etcd Put performance considerably.

Benchmark with the fix (`dbbackend=etcd`):
```
=== RUN   TestLightningNetworkDaemon
=== RUN   TestLightningNetworkDaemon/37-of-82/btcd/async_payments_benchmark
    test_harness.go:120: 	Benchmark info: Elapsed time:  6.096842747s
    test_harness.go:120: 	Benchmark info: TPS:  79.22133144038592
--- PASS: TestLightningNetworkDaemon (17.10s)
    --- PASS: TestLightningNetworkDaemon/37-of-82/btcd/async_payments_benchmark (16.42s)
```

Without the fix:
```
=== RUN   TestLightningNetworkDaemon
=== RUN   TestLightningNetworkDaemon/37-of-82/btcd/async_payments_benchmark
    test_harness.go:120: 	Benchmark info: Elapsed time:  6.222378347s
    test_harness.go:120: 	Benchmark info: TPS:  77.62305232256877
--- PASS: TestLightningNetworkDaemon (16.93s)
```